Add `LocalLifetime` and `LocalLifetimeEnd` syntax.
We add a new MIR statement, `LocalLifetimeEnd`, which indicates that
a certain local lifetime is ending. Local lifetimes are started by
calls and end at a `LocalLifetimeEnd` statement, or alternatively
somewhere in the parent body. In future, there will be a pass inserting
these statements, but this commit only changes the syntax and fixes the
resulting errors.